requirements.txt
asgiref==3.5.2
black==22.10.0
cfgv==3.3.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
filelock==3.8.2
flake8==6.0.0
identify==2.5.9
install==1.3.5
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.2
platformdirs==2.5.4
pre-commit==2.20.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
virtualenv==20.17.0
